backend/pyproject.toml
[project]
name = "rent-right-agent"
version = "0.1.0"
description = "AI-powered German tenant rights agent with Langfuse tracing"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "langfuse>=4.0.6",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
